feat: get lobby and logs links in sidekick (#235)
This is part of GDT-126
